Moved user management components into sub folder
