+++ RoomBookingApp/src/Components/UserManagementComponents/CreateUserDialog.tsx
@@ -5,14 +5,15 @@ import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogTitle from '@mui/material/DialogTitle';
-import resourceManager from "../Utilities/ResourceManager";
-import { IUser } from "../types/IUser";
+import resourceManager from "../../Utilities/ResourceManager";
+import { IconButton, MenuItem } from "@mui/material";
+import Add from "@mui/icons-material/Add";
 
-interface IChangePasswordDialog {
-    user?: IUser;
+interface ICreateUserDialog {
+    updateUsers: () => void;
 }
 
-const ChangePasswordDialog = (props: IChangePasswordDialog) => {
+const CreateUserDialog = (props: ICreateUserDialog) => {
     const [open, setOpen] = useState(false);
     const [errorMessage, setErrorMessage] = useState("");
 
@@ -26,12 +27,12 @@ const ChangePasswordDialog = (props: IChangePasswordDialog) => {
 
     const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
-        if (!props.user) return;
         setErrorMessage("");
         const formData = new FormData(event.currentTarget);
         const formJson = Object.fromEntries((formData as any).entries());
-        const request = resourceManager.makeRequest("/api/user/changePassword/" + props.user.id, "PUT", JSON.stringify(formJson), { headers: { "Content-Type": "application/json" } });
-        request.getResponse().then((response) => {
+        const request = resourceManager.makeRequest("/api/user/createUser/", "POST", JSON.stringify(formJson), { headers: { "Content-Type": "application/json" } });
+        request.getResponse().then(() => {
+            props.updateUsers();
             handleCloseDialog();
         }).catch((error) => {
             setErrorMessage(error.response.data.error_message);
@@ -40,9 +41,7 @@ const ChangePasswordDialog = (props: IChangePasswordDialog) => {
 
     return (
         <React.Fragment>
-            <Button variant="contained" onClick={handleOpenDialog}>
-                Change password
-            </Button>
+            <IconButton sx={{ width: "40px", height: "40px", marginY: "auto", backgroundColor: "#1976d2", color: "#FFFFFF" }} onClick={handleOpenDialog}><Add /></IconButton>
             <Dialog
                 open={open}
                 onClose={handleCloseDialog}
@@ -53,14 +52,14 @@ const ChangePasswordDialog = (props: IChangePasswordDialog) => {
                     },
                 }}
             >
-                <DialogTitle>Change password</DialogTitle>
+                <DialogTitle>Create user</DialogTitle>
                 <DialogContent>
                     <TextField
                         required
                         margin="dense"
-                        name="oldPassword"
-                        label="Old password"
-                        type="password"
+                        name="username"
+                        label="Name"
+                        type="text"
                         fullWidth
                         variant="standard"
                     />
@@ -68,29 +67,33 @@ const ChangePasswordDialog = (props: IChangePasswordDialog) => {
                         required
                         margin="dense"
                         name="password"
-                        label="New password"
+                        label="Password"
                         type="password"
                         fullWidth
                         variant="standard"
                     />
                     <TextField
                         required
+                        select
                         margin="dense"
-                        name="password_confirmation"
-                        label="New password (repeat)"
-                        type="password"
+                        name="role"
+                        label="Role"
                         fullWidth
                         variant="standard"
-                    />
+                        defaultValue={"teacher"}
+                    >
+                        <MenuItem value={"admin"}>Admin</MenuItem>
+                        <MenuItem value={"teacher"}>Teacher</MenuItem>
+                    </TextField>
                     {errorMessage && <span>{errorMessage}</span>}
                 </DialogContent>
                 <DialogActions>
                     <Button onClick={handleCloseDialog}>Cancel</Button>
-                    <Button type="submit">Update</Button>
+                    <Button type="submit">Create</Button>
                 </DialogActions>
             </Dialog>
         </React.Fragment>
     );
 }
 
-export default ChangePasswordDialog;
+export default CreateUserDialog;
